Adding workaround for fdopen pseudo-bug on glibc.
Currently fdopen in glibc calls fnctl to vet that the mode
bits on the handle it receives are valid in relation to the
mode arguments to fdopen. fnctl is not currently interceptable,
and instead always fails.

For now, use fopencookie to provide an alternate implementation
of fdopen.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2120
TEST=nethack works
R=jeffbailey@google.com
Review URL: https://chromiumcodereview.appspot.com/10543067