Fix openal NaCl port.
The example code was only decoding the first few openAl
buffers and then just playing the same buffer in loop.
This was due to MinBufferSizeInBytes being less than the
PPAPI buffer size which could lead to the buffer being
too empty for the reader (PPAPI) and not empty enough for
writer.

This change consists of:
 - go-between buffer size takes into account PPAPI
   buffer size as well as openAL buffer size.
 - having PPAPI callback buffer whatever samples are in
   the buffer, even if there are not enough

BUG=None
TEST=run openal-ogg example app

Review URL: https://codereview.chromium.org/12114013