buildbot: Upload naclports SDK bundle to GS storage.
Each time the SDK bundle builder runs it archives the
bundle here:
https://commondatastorage.googleapis.com/nativeclient-mirror/naclports

This allows bundles built from the head of naclports
to be downloaded and tested before the changes are
rolled into the chrome repo.

Also, rename buildbot scripts.  The only script that
changed as well as being renamed is the sdk bundle
builder.

Also, allow for fast iteration by disabling SDK download
and cleaning of the naclports build when TEST_BUILDBOT
is set.

BUG=
R=binji@chromium.org

Review URL: https://codereview.chromium.org/14585008