Allow for environment isolation in build scripts
Some applications require intermediate binaries (e.g. compilers,
interpreters, code generators) as a part of the build. Those with
support for cross-compilation assume that you set appropriate
environment variables (e.g. CC, CXX, CFLAGS, CXXFLAGS) during
configuration to point to the target system toolchain (which is
Native Client in our case), but during the actual build these variables
point the host system toolchain. Howeverm this does not work with the
current build scripts which export all variables polutting the
environment.

This patch is a step towards fixing this issue. I have moved all the
variable exports into the respective functions where they are actually
needed instead of exporting them globaly. This allows build scripts
for individual application to invoke these steps in a subshell to
achieve the environment isolation if needed.

That means that the following function:

ConfigureStep() {
  ...
  DefaultConfigureStep
  ...
}

would become:

ConfigureStep() {
  ...
  (DefaultConfigureStep)
  ...
}

ensuring that any variables exported inside the DefaultConfigureStep
do not escape and polute the environment. For now, this is opt-in,
which means that the steps which should be isolated have to manually
wrapped in a subshell. If this proves to be useful, we might decide
to wrap all steps in a subshell by default in the future.

R=bradnelson@google.com

Review URL: https://codereview.chromium.org/23654053