Fix toybox build for getcwd, nacl_spawn and configure.
As noted in https://codereview.chromium.org/140993003/ , the behavior
for getcwd in case of NULL or 0 size should be controlled by the libc
rather than the KernelProxy.  We now allocate a buffer on the heap and
pass that in to the libc getcwd.  Furthermore, the functions defined
inside the portability routine for operating on dirfd arguments are not
yet defined in the GLIBC implementation, and so we should define them
here.

This also fixes an issue with nacl_waitpid and a mandatory persistent
html5fs, where combined to cause crashes with the shipped html.

The final modification is to manually specify the configure path for the
build process as including sh.

R=sbc@chromium.org
BUG=https://code.google.com/p/naclports/issues/detail?id=104

Review URL: https://codereview.chromium.org/140993005