buildbot: Add missing bzip2 to partition 0 of 3
This dependency was added to freetype by a previous commit

Also, fix build_shard.sh so that is can handle failures
in partition.py in a sane way.

Also update BUILD_ROOT in naclports.py (it was still pointing
to the old 'repository' location).

R=binji@chromium.org

Review URL: https://codereview.chromium.org/211383005